Remove old survey admin code (actions, reducers, components), use survey route for new version
