+++ assembl/static2/js/app/pages/administration.jsx
@@ -6,7 +6,7 @@ import { Grid, Row, Col } from 'react-bootstrap';
 import { connect } from 'react-redux';
 import moment from 'moment';
 
-import { updateThematics, displayLanguageMenu } from '../actions/adminActions';
+import { displayLanguageMenu } from '../actions/adminActions';
 import { updateResources, updateResourcesCenterPage } from '../actions/adminActions/resourcesCenter';
 import { updateVoteSessionPage, updateVoteModules, updateVoteProposals } from '../actions/adminActions/voteSession';
 import { updatePhases } from '../actions/adminActions/timeline';
@@ -17,7 +17,6 @@ import { updateTextFields } from '../actions/adminActions/profileOptions';
 import withLoadingIndicator from '../components/common/withLoadingIndicator';
 import Menu from '../components/administration/menu';
 import LanguageMenu from '../components/administration/languageMenu';
-import ThematicsQuery from '../graphql/ThematicsQuery.graphql';
 import ResourcesQuery from '../graphql/ResourcesQuery.graphql';
 import ResourcesCenterPage from '../graphql/ResourcesCenterPage.graphql';
 import SectionsQuery from '../graphql/SectionsQuery.graphql';
@@ -31,35 +30,10 @@ import { convertEntriesToRawContentState } from '../utils/draftjs';
 import { getPhaseId } from '../utils/timeline';
 import landingPagePlugin from '../utils/administration/landingPage';
 
-export function convertVideoDescriptions(thematics) {
-  return thematics.map((t) => {
-    if (!t.video) {
-      return t;
-    }
-
-    return {
-      ...t,
-      video: {
-        ...t.video,
-        descriptionEntriesBottom: t.video.descriptionEntriesBottom
-          ? convertEntriesToRawContentState(t.video.descriptionEntriesBottom)
-          : null,
-        descriptionEntriesSide: t.video.descriptionEntriesSide
-          ? convertEntriesToRawContentState(t.video.descriptionEntriesSide)
-          : null,
-        descriptionEntriesTop: t.video.descriptionEntriesTop
-          ? convertEntriesToRawContentState(t.video.descriptionEntriesTop)
-          : null
-      }
-    };
-  });
-}
-
 class Administration extends React.Component {
   constructor(props) {
     super(props);
     this.putResourcesCenterInStore = this.putResourcesCenterInStore.bind(this);
-    this.putThematicsInStore = this.putThematicsInStore.bind(this);
     this.putLegalContentsInStore = this.putLegalContentsInStore.bind(this);
     this.putVoteSessionInStore = this.putVoteSessionInStore.bind(this);
     this.putLandingPageModulesInStore = this.putLandingPageModulesInStore.bind(this);
@@ -75,7 +49,6 @@ class Administration extends React.Component {
     // "global" save button that will do all the mutations "at once"
     this.putResourcesCenterInStore(this.props.resourcesCenter);
     this.putResourcesInStore(this.props.resources);
-    this.putThematicsInStore(this.props.data);
     this.putSectionsInStore(this.props.sections);
     this.putLegalContentsInStore(this.props.legalContents);
     this.putVoteSessionInStore(this.props.voteSession);
@@ -90,11 +63,6 @@ class Administration extends React.Component {
   }
 
   componentWillReceiveProps(nextProps) {
-    // update thematics in store after a mutation has been executed
-    if (nextProps.data.thematics !== this.props.data.thematics) {
-      this.putThematicsInStore(nextProps.data);
-    }
-
     if (nextProps.resources !== this.props.resources) {
       this.putResourcesInStore(nextProps.resources);
     }
@@ -133,13 +101,6 @@ class Administration extends React.Component {
     if (nextProps.timeline !== this.props.timeline) {
       this.putTimelinePhasesInStore(nextProps.timeline);
     }
-  }
-
-  putThematicsInStore(data) {
-    // filter with the same query to remove stuff like __typename from the structure
-    const filteredThematics = filter(ThematicsQuery, data);
-    const thematics = convertVideoDescriptions(filteredThematics.thematics);
-    this.props.updateThematics(thematics);
   }
 
   putResourcesInStore(resources) {
@@ -262,7 +223,6 @@ class Administration extends React.Component {
   render() {
     const {
       children,
-      data,
       i18n,
       params,
       refetchResources,
@@ -282,7 +242,6 @@ class Administration extends React.Component {
       React.cloneElement(child, {
         locale: i18n.locale,
         refetchTabsConditions: refetchTabsConditions,
-        refetchThematics: data.refetch,
         refetchResources: refetchResources,
         refetchVoteSession: refetchVoteSession,
         refetchSections: refetchSections,
@@ -345,7 +304,6 @@ const mapStateToProps = state => ({
 const mapDispatchToProps = dispatch => ({
   updateResources: resources => dispatch(updateResources(resources)),
   updateSections: sections => dispatch(updateSections(sections)),
-  updateThematics: thematics => dispatch(updateThematics(thematics)),
   updateResourcesCenterPage: ({ titleEntries, headerImage }) => {
     dispatch(updateResourcesCenterPage(titleEntries, headerImage));
   },
@@ -362,7 +320,6 @@ const mapDispatchToProps = dispatch => ({
 
 const mergeLoadingAndHasErrors = WrappedComponent => (props) => {
   const {
-    data,
     resourcesHasErrors,
     resourcesCenterHasErrors,
     resourcesLoading,
@@ -393,8 +350,7 @@ const mergeLoadingAndHasErrors = WrappedComponent => (props) => {
     sectionsHasErrors ||
     props[landingPagePlugin.hasErrors] ||
     textFieldsHasErrors ||
-    timelineHasErrors ||
-    (data && data.error);
+    timelineHasErrors;
   const loading =
     voteSessionLoading ||
     resourcesLoading ||
@@ -405,17 +361,13 @@ const mergeLoadingAndHasErrors = WrappedComponent => (props) => {
     sectionsLoading ||
     props[landingPagePlugin.loading] ||
     textFieldsLoading ||
-    timelineIsLoading ||
-    (data && data.loading);
+    timelineIsLoading;
 
   return <WrappedComponent {...props} hasErrors={hasErrors} loading={loading} />;
 };
 
 export default compose(
   connect(mapStateToProps, mapDispatchToProps),
-  graphql(ThematicsQuery, {
-    options: { variables: { identifier: 'survey' } }
-  }),
   graphql(TabsConditionQuery, {
     options: ({ i18n }) => ({
       variables: { lang: i18n.locale }
